# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BlueCore SMM Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a lead (consultation request)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    get:
      operationId: getLeads
      tags: [leads]
      summary: Get leads list (admin/manager only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: service
          in: query
          schema:
            type: string
        - name: lang
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsResponse"
  /leads/export/csv:
    get:
      operationId: exportLeadsCsv
      tags: [leads]
      summary: Export leads as CSV (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a single lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status or assignee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                assignedTo:
                  type: number
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/notes:
    post:
      operationId: addLeadNote
      tags: [leads]
      summary: Add a note to a lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
              required: [text]
      responses:
        "200":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get all active services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{slug}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send notification to user(s) (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked as read
...
  /translations/{lang}/{namespace}:
    get:
      operationId: getTranslations
      tags: [translations]
      summary: Get translations for a namespace
      parameters:
        - name: lang
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Translations object
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /translations/admin/all:
    get:
      operationId: getAllTranslations
      tags: [translations]
      summary: Get all translations (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Translation"
  /translations:
    put:
      operationId: upsertTranslation
      tags: [translations]
      summary: Create or update a translation (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationInput"
      responses:
        "200":
          description: Translation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Translation"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateLeadRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        company:
          type: string
        service:
          type: string
        message:
          type: string
        source:
          type: string
        lang:
          type: string
      required: [name, phone]
    Lead:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        company:
          type: string
        service:
          type: string
        message:
          type: string
        source:
          type: string
        status:
          type: string
        assignedTo:
          type: number
        notes:
          type: array
          items:
            type: object
        ip:
          type: string
        lang:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, phone, status, lang]
    LeadsResponse:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [leads, total, page, limit]
    MultiLangContent:
      type: object
      properties:
        uz:
          type: object
        ru:
          type: object
        en:
          type: object
...
    ServiceInput:
      type: object
      properties:
        slug:
          type: string
        icon:
          type: string
        order:
          type: number
        active:
          type: boolean
        content:
          $ref: "#/components/schemas/MultiLangContent"
        pricing:
          type: object
      required: [slug, content]
    Service:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        icon:
          type: string
        order:
          type: number
        active:
          type: boolean
        content:
          $ref: "#/components/schemas/MultiLangContent"
        pricing:
          type: object
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, slug, active, content]
...
    NotificationInput:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        type:
          type: string
        link:
          type: string
        userId:
          type: number
      required: [title, body]
    Notification:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        link:
          type: string
        createdAt:
          type: string
      required: [id, type, title, body, read]
...
    TranslationInput:
      type: object
      properties:
        namespace:
          type: string
        key:
          type: string
        uz:
          type: string
        ru:
          type: string
        en:
          type: string
      required: [namespace, key]
    Translation:
      type: object
      properties:
        id:
          type: number
        namespace:
          type: string
        key:
          type: string
        uz:
          type: string
        ru:
          type: string
        en:
          type: string
        updatedAt:
          type: string
      required: [id, namespace, key]