# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueCore SMM Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { MultiLangContent } from "./multiLangContent";
import type { ServicePricing } from "./servicePricing";

export interface Service {
  id: number;
  slug: string;
  icon?: string;
  order?: number;
  active: boolean;
  content: MultiLangContent;
  pricing?: ServicePricing;
  createdAt?: string;
  updatedAt?: string;
}